pyproject.toml
[project]
name = "trading-notification-bot"
version = "1.0.0"
description = "Automated bot that fetches market data and posts to X (Twitter) and Discord"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "tweepy>=4.14.0",
    "discord-webhook>=1.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pillow>=10.2.0",
    "loguru>=0.7.2",
    "apscheduler>=3.10.4",
    "pytz>=2024.1",
    "aiohttp>=3.9.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
